+++ src/LinkDotNet.StringBuilder/ValueStringBuilder.AppendJoin.cs
@@ -42,6 +42,7 @@ public ref partial struct ValueStringBuilder
     public void AppendJoin<T>(char separator, IEnumerable<T> values)
         => AppendJoinInternalChar(separator, values);
 
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private void AppendJoinInternalString<T2>(ReadOnlySpan<char> separator, IEnumerable<T2> values)
     {
         ArgumentNullException.ThrowIfNull(values);
@@ -70,6 +71,7 @@ public ref partial struct ValueStringBuilder
         }
     }
 
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private void AppendJoinInternalChar<T2>(char separator, IEnumerable<T2> values)
     {
         ArgumentNullException.ThrowIfNull(values);
@@ -98,6 +100,7 @@ public ref partial struct ValueStringBuilder
         }
     }
 
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private void AppendInternal<T>(T value)
     {
         if (value is ISpanFormattable spanFormattable)